pyproject.toml
[tool.poetry]
name = "trafficgenerator"
version = "0.1.0"
description = ""
authors = ["ridabourji <70974349+ridabourji@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pysocks = "^1.7.1"
stem = "^1.8.2"
playwright = "^1.47.0"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
